Rubik's Cube Mathematica
This is an attempt to dissect this demonstration code .  This demonstration is hosted at:
http://demonstrations.wolfram.com/RubiksCube/
it consists of an animation of the twisting of a Rubik's Cube. Use the six colored buttons to cause a clockwise quarter-turn twist of the face with the corresponding color at the center. Holding down the shift key causes a counterclockwise turn instead. Use the control key to produce a half turn. The slider controls the number of steps of the animation of each twist.
Contributed by: Roman E. Maeder